+++ Tests/Editor/LoggerTest.cs
@@ -14,6 +14,7 @@
 //
 
 using System;
+using System.Linq;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -27,9 +28,9 @@ namespace GLTFast.Tests
             var r = new CollectingLogger();
             r.Error(LogCode.Download,"404", "https://something.com/nowherfound.glb");
             
-            Assert.NotNull(r.items);
-            Assert.AreEqual(1,r.items.Count);
-            Assert.AreEqual("Download URL https://something.com/nowherfound.glb failed: 404", r.items[0].ToString());
+            Assert.AreEqual(1,r.Count);
+            var items = r.Items.ToArray();
+            Assert.AreEqual("Download URL https://something.com/nowherfound.glb failed: 404", items[0].ToString());
         }
         
         [Test]