+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.0.0-pre]
 ### Changed
 - Converted a lot of unintentionally public classes, types and properties to internal ones
+- Replaced `CollectingLogger.item` with `.Count` and `.Items` iterator
 
 ## [Unreleased]
 ### Added